ENH: added `load_radar_site` routine
Added the `load_radar_site` routine to the radar library.  Updated make_fov and make_grid.2.0 routines to use this new function.
